Change seedu.duke to seedu.moviemate. Delete Duke.java, DukeTest.java. Change build.gradle.
