+++ META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: MovieMate
+Main-Class: seedu.moviemate.MovieMate
