+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass = "seedu.duke.Duke"
+    mainClass = "seedu.moviemate.MovieMate"
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "moviemate"
     archiveClassifier = null
 }
