--- src/main/java/HelpCommand.java
@@ -1,8 +1,0 @@
-
-public class HelpCommand extends Command {
-
-    @Override
-    public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
-        ui.printHelp();
-    }
-}
