+++ src/main/java/seedu/moviemate/AddWatchedListCommand.java
@@ -1,14 +1,16 @@
+package seedu.moviemate;
 
-public class AddToWatchListCommand extends Command {
+public class AddWatchedListCommand extends Command {
+
     private String inputTitle;
 
-    public AddToWatchListCommand(String inputTitle) {
+    public AddWatchedListCommand(String inputTitle) {
         this.inputTitle = inputTitle;
     }
 
     @Override
     public  void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
-        // add to to-watch list
-        toWatchList.add(inputTitle);
+        // add to watched list
+        watchedList.add(inputTitle);
     }
 }
