+++ src/main/java/seedu/moviemate/Command.java
@@ -1,4 +1,4 @@
-
+package seedu.moviemate;
 
 public abstract class Command {
     public abstract void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage);
