+++ src/main/java/seedu/moviemate/ExitCommand.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 public class ExitCommand extends Command {
 
     @Override
