+++ src/main/java/seedu/moviemate/HelpCommand.java
@@ -1,8 +1,9 @@
-public class WatchListCommand extends Command {
+package seedu.moviemate;
+
+public class HelpCommand extends Command {
 
     @Override
     public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
-        // list the towatch list
-        toWatchList.list();
+        ui.printHelp();
     }
 }
