+++ src/main/java/seedu/moviemate/ListCommand.java
@@ -1,3 +1,4 @@
+package seedu.moviemate;
 
 public class ListCommand extends Command {
