+++ src/main/java/seedu/moviemate/MovieList.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
