+++ src/main/java/seedu/moviemate/MovieMate.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
