+++ src/main/java/seedu/moviemate/Parser.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 /**
  * Parses user input.
  */
