+++ src/main/java/seedu/moviemate/ReadCSVFile.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
