+++ src/main/java/seedu/moviemate/Storage.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
