+++ src/main/java/seedu/moviemate/ToWatchList.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 import java.util.ArrayList;
 
 /**
