+++ src/main/java/seedu/moviemate/Ui.java
@@ -1,3 +1,5 @@
+package seedu.moviemate;
+
 import java.util.Scanner;
 
 /**
