+++ src/main/java/seedu/moviemate/ViewReviewCommand.java
@@ -1,3 +1,4 @@
+package seedu.moviemate;
 
 public class ViewReviewCommand extends Command {
