+++ src/main/java/seedu/moviemate/WatchListCommand.java
@@ -1,0 +1,10 @@
+package seedu.moviemate;
+
+public class WatchListCommand extends Command {
+
+    @Override
+    public void execute(WatchedList watchedList, ToWatchList toWatchList, Ui ui, Storage storage) {
+        // list the towatch list
+        toWatchList.list();
+    }
+}
