+++ src/test/java/seedu/moviemate/DukeTest.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.moviemate;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
